+++ Machine.Specifications.ReSharper.Tests/Psi/ContextTests.cs
@@ -52,13 +52,7 @@ namespace Machine.Specifications.ReSharper.Tests.Psi
         [Test]
         public void GenericClassIsNotAContext()
         {
-            WithPsiFile("GenericWithSpecs.cs", x =>
-            {
-                var f = Fields.FirstOrDefault().AsFieldInfo();
-                f.FieldType.GetGenericArguments();
-
-                Assert.That(Type().IsContext(), Is.False);
-            });
+            WithPsiFile("GenericWithSpecs.cs", x => Assert.That(Type().IsContext(), Is.False));
         }
     }
 }
